+++ sandbox/rules.h
@@ -1,7 +1,7 @@
 #ifndef RULES_H
 #define RULES_H
-uintmax_t RuleNr(const int, const int, const std::vector<int>);
-std::vector<int>* RuleParts(const int, const int, const uintmax_t);
+long long unsigned RuleNr(const int, const int, const std::vector<int>);
+std::vector<int>* RuleParts(const int, const int, const long long unsigned);
 
 // Action numbers
 #define AWHITE 0
